h1load: respect the global limit on the request count
Now if there's a limit on the global request count, it's checked before
sending a new request. This is done atomically so that all threads agree
on which request may be sent. It's also used to stop creating new
connections and to stop the master thread's waiting loop. Threads which
notice the limit is reached when sending their last requests will also
emit connection:close on the respective connections. This is mainly to
ensure that at low loads and/or on single connections, the finishing is
clean in network traces.