h1load: add support for protocol overhead correction
This is important to measure traffic at the TCP/IP/ethernet level. One
just has to pass the number of added bytes per payload bytes. Typically
one may pass:

  -O 20/1460  to count the TCP layer
  -O 40/1460  to count the IP layer
  -O 54/1460  to count the Ethernet L2 layer
  -O 78/1460  to count the Ethernet on-wire layer

Add 12 of overhead and subtract 12 of payload when using TCP timestamps,
thus for on-wire Ethernet with TCP timestamps, it gives: -O 90/1448