h1load: always support creating twice as more connections as events
With high loads of non-persistent connections sometimes it's impossible
to reach the desired amount of connections because we don't create more
than we can kill in one loop. This patch changes this by allowing to
create up to twice the number of events in terms of connections. This
allows them to start much faster now.