h1load: implement chunked encoded responses
Till now they were handled as a tunnel, but now this provides a simple
implementation. It assumes that messages are relatively correctly
formatted (i.e. CRLF is assumed after the data but any two bytes will
work). It's not the fastest in the world with large messages but with
those that fit into the recv buffer it manages to parse rougly 740
million chunks per second or 36 Gbps of minimally chunked data on
2 cores/4 threads.

Note that trailers are not supported, for now the message is assumed
to end with 0 CRLF CRLF to avoid saving one extra state.