h1load: add new option -P to dump percentiles of ttfb/ttlb
The percentiles are dumped with various values to help graphing them,
such as the percentile, the tail (1-pct), and the inverse of the tail,
which allows to use log scales in gnuplot. The counts for each set are
reported, as well as the values in milliseconds.

In order to keep things fast, a type of 16-bit float was created.
This has the benefit of only requiring 2^16 counters for a range of
2^42, and of keeping entries ordered to ease the dump.

The values can be plotted in gnuplot by taking the labels from the
first column like this:

    plot "file" using 3:5:xticlabel(1) with lines

This however tends to emit many tick labels. A working alternative
consists in adding fixed labels like this:

 set label  1 "10%" at   1.11111,-0.5 center front
 set label  3 "30%" at   1.42857,-0.5 center front
 set label  5 "50%" at   2.00000,-0.5 center front
 set label  7 "70%" at   3.33333,-0.5 center front
 set label  8 "80%" at   5,-0.5 center front
 set label  9 "90%" at  10,-0.5 center front
 set label 10 "95%" at  20,-0.5 center front
 set label 11 "97.5%" at  40,-0.5 center front
 set label 12 "99%" at 100,-0.5 center front
 set label 13 "99.7%" at 333,-0.5 center front
 set label 15 "99.9%" at 1000,-0.5 center front
 set label 16 "99.97%" at 3333,-0.5 center front
 set label 18 "99.99%" at 10000,-0.5 center front
 set label 19 "99.997%" at 33333,-0.5 center front
 set label 21 "99.999%" at 100000,-0.5 center front
 set label 22 "99.9995%" at 200000,-0.5 center front
 set label 23 "100%" at 250000,-0.5 center front

Or these may be placed into a 3-column file plotted using the
"with labels" mode.