h1load: make the request counter more accurate using an extra state
In the very unlikely case where we'd fail to emit the request due to
extreme pressure in the network buffers, we would have counted a request
twice because the I/O and the request are performed together. Better have
a CS_REQ state to count +1 on the request and let CS_SND form the request
and emit it.