h1load: count done requests instead of started ones
One issue emphasized by the previous fix is that we used to count started
requests instead of completed ones. The effect is that the first second
tends to show one extra cycle of requests and that the last one usually
shows none. This is inaccurate as it doesn't reflect the server's ability
to respond, only the client's ability to produce requests.

This patch changes this by only reporting completed requests (either failed
or successful). This is way more accurate. Connections still do not count
for a request, so that only successful connections will try to emit a
request.

Now a test against httpterm properly reports:

  $ ./h1load -e -n 30 -c 10  0:8000/?t=1000
  #     time conns tot_conn  tot_req      tot_bytes    err  cps  rps  Bps  bps
  1612783410    10       10       10           1320      0 10.0 10.0 1k32 10k5
  1612783411    10       10       20           2640      0 0.00 9.99 1k31 10k5
  1612783412     0       10       30           3979      0 0.00 9.99 1k33 10k7

With multiple threads:

  $ ./h1load -t 4 -e -n 32 -c 16  0:8000/?t=1000
  #     time conns tot_conn  tot_req      tot_bytes    err  cps  rps  Bps  bps
  1612783459    16       16       10           1320      0 16.0 10.0 1k32 10k5
  1612783460     7       16       25           3300      0 0.00 14.9 1k97 15k8
  1612783461     0       16       32           4243      0 0.00 6.99 942. 7k53

Then smoothed with a slowstart:

  $ ./h1load -t 4 -e -n 64 -s 1000 -c 16  0:8000/?t=1000
  #     time conns tot_conn  tot_req      tot_bytes    err  cps  rps  Bps  bps
  1612783535     4        4        1            132      0 4.00 1.00 132. 1k05
  1612783536    16       16        5            660      0 11.9 3.99 527. 4k21
  1612783537    16       16       21           2772      0 0.00 15.9 2k10 16k8
  1612783538    16       16       40           5280      0 0.00 18.9 2k50 20k0
  1612783539     5       16       59           7788      0 0.00 18.9 2k50 20k0
  1612783540     0       16       64           8467      0 0.00 4.99 678. 5k42