h1load: differentiate soft-stopping from hard-stopping
Right now there are always a few requests at the end that are aborted
when a limited request count is set. The reason for this is that the main
thread exit the loop as soon as the total request count is reached, and
sends a hard-stop flag to all threads, which can't even complete their
polling loop.

Let's change the way this is done. A new flag THR_ENDING indicates that
threads must gracefully terminate what they were doing, refrain from
emitting new requests, abandon idle connections, and exit. The main
thread now uses this flag to end on duration or request count. And
instead of quitting its loop, the main thread now waits for other
threads to be gone, which allows to continue to periodically report
their status instead of waiting blindly.

The difference is that a duration-limited test may now last a bit longer
for the time it takes to complete pending requests, especially if some
network errors are encountered. But this is not the most common case, and
it remains more desirable than the current one where successful responses
are not handled at the end.

Now the hard stop will only be used to force threads to immediately exit,
which the main thread will notice and do the same.

However, this alone is not sufficient and emphasizes a race that happens
on startup where some threads could finish their job while others are
waiting for all of them to start, and the main thread would never see
them all ready. In order to address this, yet another flag, THR_SYNSTART
is set while all threads are waiting for each other, so that they don't
have to watch each other, and the main thread is the only one counting
them and clearing that flag to let them go.

As a bonus, it should now be much easier to detect startup errors earlier
instead of waiting for a full sleep().