h1load: make the request rate follow the slow start throttling as well
The request rate limitation wasn't subject to the slow start, which used
to only apply to the concurrent connections. Now this also follows the
throttling.

This is far from being trivial because new connections compete with new
requests over the same connection, and this is a problem when raising
the load, due to an excess of queued requests that make the frequency
counters report extremely high wait times. The main issue is that once
the request rate is achieved on a small number of connections, there's
no more place for creating new connections.

Thus we have to limit the drift to two seconds, and to let connections
continue to establish. We also make sure to always leave some budget
for new connections if the request rate is non-nul. The ramp-up might
be a bit chaotic in certain circumstances but overall it's OK.