h1load: implement a slightly better still simple sleep queue
We have few time variations in the sleep queue due to the modulation of
the think time during the ramp up. Since we used to stop on the first
non-expired entry in the list, this started to becomme unstable. It
still doesn't seem needed to switch to a fully ordered queue at this
point, given that the time offsets remain small.

What this patch does is to implement 32 sleep queues per thread, and
tasks are put to sleep there based on their timer bits 17..21. This
is sufficient to keep them reasonably ordered modulo 125 ms across a
range of 4 seconds, and is sufficient to avoid large holes and peaks
during the ramp up.