h1load: make the slowstart rampups smoother
We used to update the throttling value every 100ms but this it too slow
for high request rate traffic as it can still require to start a huge
amount of connections on each step. Let's decrease this to 20ms which
remains a multiple of 4 and 10ms which are commonly found as system
timers.