h1load: use a cleaner soft-stopping state than hacking the arg_pctl
Soft stopping was still dirty because we were not completely stopping
measurements when the delay was over due to the hack around the
percentile. Let's add a new flag THR_DUR_OVER to indicate that the
test is over, and use this to synchronously stop collecting values
before threads exit. This slightly changes the main loop, and this
also makes sure that we now wait for the last requests to be finished
when working on request counts.