h1load: improve thread consistency in periodic reports
While for most columns we don't really care how recent the value is,
it is important that the tot_ttfb count is synchronized with tot_fbs
and that tot_ttlb is synchronized with tot_lbs, otherwise we can
occasionally report erratic values because one is divided by the other.
And this happens because at least one test showed average times much
larger than a second in the first second due to too low a number of
samples. Let's just use a few atomic stores there, and load all values
using atomic loads.

In theory the other values should be stored using an atomic_store but
in practice they're flushed fast enough so we don't really care.