h1load: add a new option "-A" for more accurate response time measurement
In some environments, establishing a connection via a NAT layer may
artificially damage the first request's response time, especially if
some SYN proxy mechanisms are involved, or if a transparent proxy is
installed for example. In such a case it becomes difficult to accurately
measure timers while they're stable after all connections are set up.

This patch brings a new option "-A" which enables more accurate response
time measurements, simply by not counting the first request of a
connection, as it might have been blocked in transit somewhere in the
chain. This must obviously not be used when running in http-close mode.