h1load: try to pre-heat the threads by allocating the required stuff
During the initial ramp-up, response times are horrible and this happens
mostly around the socket() call, but not only. Times of 35ms and more
were observed on various machines, likely due to allocations in the system
and maybe even in the libc. Other long times happen in connect().

The problem is that this causes erratic measurements at the beginning of
a test, as well as during all the ramp up. This is mostly noticeable at
moderate loads and moderate numbers of connections. Most of the time this
is not even noticeable below 250 connections, as if there were some FD
ranges to allocate somewhere. But running with -R 500 -c 1000 definitely
exhibits the problem. And this problem is solely a client issue, as it
never re-appears during the load but immediately re-appears if the client
is stopped and restarted. It's very likely related to some internal
locking of the FD table or so, as established connections are also
affected.

In order to partially address this, this patch adds a pre-heating phase
before all threads start, during which they allocate the max number of
sockets and connections they'll ever have to support. This also allows
to quickly detect early allocation failures. The starting phase now
happens in two steps, allocation and closure, both finishing when
running==arg_thrd.

Doing this has reliably reduced the first second ttfb from ~4-5 ms to
more like 1.5 ms, which is still far from being acceptable but at
least one part of the issue is addressed.