scripts: add more flexible, graphing scripts
These new scripts process one file at a time but support command line
options to adjust their behavior or to send the output to a specific
file (no write permissions needed anymore in the data directory). They
can optionally take a reference file or not. The 3 scripts vary by the
way they handle the injected load: not reported, reported stacked, or
reported on a separate graph.