scripts: make graph_pct support an optional second file and individual titles
Now it's possible to compare a value against the reference by passing the two
files in argument. Each name may be suffixed with ":" followed by a title to
display on the key box. A log scale is supported, the max value(s) appear on
the key, and the title will automatically be adjusted to "x vs y" in case two
entries are passed.