h1load: implement ssl at the connection level using the fd BIO
With this it now works. If the connection is flagged as using SSL, it
allocates an SSL object and uses SSL_read() / SSL_write() at the fd
level. This is ugly under the hood as it uses read()/write() instead
of the cleaner recv()/send() and is why we need to disable SIGPIPE.
But overall it seems to work and no data appears to block inside the
buffers since we only set the polling flags on WANT_READ/WANT_WRITE.

We do not set the new_cb callback so we never resume connections in
case of close. In practice it's not really an issue as this usually
is what is desired in a benchmark utility.