h1load: use a single destination for the pre-heat phase
Calling bind() on each temporary connection is extremely expensive, and
starting a single thread with 30k connections when the limit is 28k takes
~10s to fail.

We don't care about the destination provided that it exists, so this patch
changes this so that the first bound connection on each thread is used as
the target for all subsequent connections. Now the startup is instantaneous.