h1load: add a handshake phase
The connect() over SSL is a bit nasty because we don't have a special
state for it. As such the SSL_connect() call is expected to immediately
succeed, but it's called right after connect() and may not necessarily
have enough TCP window to send everything if the round trip is not
complete, in which case it will be handled by the next recv()/send()
call, or conversely might succeed without telling the poller, leading
the poller to still poll for writes after the last write was done.
That's neither efficient nor clean, and may fail in EPOLLET mode. In
addition, where we are we cannot check the return of SSL_connect() so
we don't know what side to poll for!

Let's have a handshake state after the connect succeeds, in which the
SSL part will be handled.