h1load: detect when too many arguments are passed to the command line
A few times already we've got caught passing certain args after the URL,
that were silently ignored. That's really bad because it can require that
a whole series of tests have to be done again. Let's add a check for the
presence of such post-URL args and refuse to start in this case.