h1load: simplify the C-L vs chunk size decounting
The loop had become bogus with the introduction of chunk size because
it was a bit more complicated. Let's simplify it so that we just need
to re-evaluate the need to quit at a few points, and ignore the rest
of the loop for non-chunk. That even removes one if() level, slightly
reduces the code and makes it 0.5% faster when parsing small chunks.