h1load: do not report too many errors in close mode
Some servers tend to close with RST and cause trouble in close mode
because we still report the error past the end of data or when in
tunnel mode while it's the only indication that the end was reached.
Sometimes it may just be because they haven't completely consumed the
request. Let's just take care of these special cases which are rare
but confusing when reported as errors.