Ensure the Host header does not contain a query string
If a user provides a URL without a path but with a query string (for
example, http://192.0.2.1:8080?s=64) the query string is ignored, and
the Host header is incorrectly set to 192.0.2.1:8080?s=64, which is
invalid.

This behavior can lead to inaccurate benchmarking results.

This patch adds a check to ensure that when no explicit path is
specified in the URL, the Host header does not include a query
string (i.e., does not contain a ?). If such a case is detected,
h1load will report an error.

Signed-off-by: Sébastien Gross <seb•ɑƬ•chezwam•ɖɵʈ•org>