h1load: report the averaged throttling ratio over the period
It's misleading and confusing to see the throttling ratio at the moment
the line is printed with -ll, because the last measure affected by the
ram-up displays "100" while it was at least partially impacted by the
throttling. Now we keep a copy of the previous measure and report the
average over the period. This is much more accurate and it allows to
detect the first measure of the plateau.