h1load: automatically limit thread count to connection count
It's common to run with "-t $(nproc)", but with low numbers of connections
this triggers an error and requires the user to manually set the value.
It's even more annoying in scripts. But there's no technical reason that
forces this to be an error, it can automatically be adjusted. This commit
does this. When run in verbose mode it will also print a warning indicating
it.